+++ H2Projekt.Application/Dto/Rooms/RoomDiscountDto.cs
@@ -9,7 +9,7 @@ namespace H2Projekt.Application.Dto.Rooms
         public string Description { get; set; }
         public DateOnly FromDate { get; set; }
         public DateOnly ToDate { get; set; }
-        public int PricePerNight { get; set; }
+        public decimal PricePerNight { get; set; }
 
         public RoomDiscountDto(RoomDiscount roomDiscount)
         {
